pyproject.toml
[tool.poetry]
name = "draft_deck"
version = "0.1.0"
description = ""
authors = ["array_fish <uekusa_hideaki@yahoo.co.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.2.0"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
